Fix naija not moving properly on machines with very high FPS.
This bug occured when standing still and attempting to move,
having high FPS and very low difftime.
With near-zero vel, it was not built up enough per frame,
therefore velLen was always < 2 and was subsequently reset back to 0,
which resulted in naija never actually moving.
This bug could be reproduced by holding Shift+F with 60 FPS, and then
holding an arrow key.

Thanks to `Nax for reporting the problem.

(Cleaned out some cruft too, the actual change is the last line of the diff)