Now Visual Studio 2010 Compatible
Removed a "using namepsace std;" to stop an ambiguous symbol error
regarding the "hash" datatype when compiling with Visual Studio 2010. As
a result, I had to add a "std::" to an ostringstream.

Duplicated, upgraded, and reworked the existing Visual Studio 2008
project so that it now compiles successfully with Visual Studio 2010.
Both debug and release work without a hitch on Windows 7. The original
Visual Studio 2008 solution is still there if it is needed.