fix long-standing editor bug that would cause entities to be moved around
This was caused by early deletion of entities.
If new ones were spawned upon further editing of a map, it could re-use
entity IDs and cause both new and old entities to be spawned in the position
of the new entity (due to EntitySaveData of the old one still intact).

This is now prevented by not touching the ID range of map entities.

Should probably make an offline tool to recompress entity IDs...