Fix editor single-axis scaling sliding the tile around
This re-implements movement compensation previoulsy done by
RenderObject::beforeScaleOffset, that got removed in 048a787cb1faaeeac0f.

Turns out all that needs to be done is to take tile rotation into account.

(The removed FIXME comment has nothing to do with it; it's just a leftover
that turned out to work fine as-is)