remove RenderObjectLayer::update. now all layers are updated, always.
The original use case of the update flag was to prevent those layers that
only had map Element tiles on them to update each tile, which would be
very slow and most tiles never had any reason to be updated.
(Those that needed to were tracked in a separate list.)

Since the map Element class does no longer exist and tiles are handled
much more efficiently now, it makes sense to not arbitarily disable
object updates for layers that used to have only map tiles on them.

Note to self: This fixes the skeleton entities properly reacting to
shots now when placed on tile layers.